Add override to unicorn config

+++ config/unicorn/production.rb
@@ -1,32 +1,36 @@
-rails_env = ENV['::Rails.env'] || 'production'
+if File.exists?("local/config/unicorn/production.rb")
+  load "local/config/unicorn/production.rb" 
+else
+  rails_env = ENV['::Rails.env'] || 'production'
 
-worker_processes 2
-preload_app true
-timeout 180
+  worker_processes 2
+  preload_app true
+  timeout 180
 
-user 'app', 'rails'
+  user 'app', 'rails'
 
-app_path = File.expand_path(File.dirname(__FILE__) + "/../../../current")
-working_directory app_path
-listen "#{app_path}/tmp/sockets/unicorn.sock", :backlog => 64
-stderr_path "log/unicorn.stderr.log"
-stdout_path "log/unicorn.stdout.log"
+  app_path = File.expand_path(File.dirname(__FILE__) + "/../../../current")
+  working_directory app_path
+  listen "#{app_path}/tmp/sockets/unicorn.sock", :backlog => 64
+  stderr_path "log/unicorn.stderr.log"
+  stdout_path "log/unicorn.stdout.log"
 
-pid "#{app_path}/tmp/pids/unicorn.pid"
+  pid "#{app_path}/tmp/pids/unicorn.pid"
 
-before_fork do |server, worker|
-  ActiveRecord::Base.connection.disconnect!
+  before_fork do |server, worker|
+    ActiveRecord::Base.connection.disconnect!
 
-  old_pid = "#{server.config[:pid]}.oldbin"
-  if File.exists?(old_pid) && server.pid != old_pid
-    begin
-      Process.kill("QUIT", File.read(old_pid).to_i)
-    rescue Errno::ENOENT, Errno::ESRCH
-      # someone else did our job for us
+    old_pid = "#{server.config[:pid]}.oldbin"
+    if File.exists?(old_pid) && server.pid != old_pid
+      begin
+        Process.kill("QUIT", File.read(old_pid).to_i)
+      rescue Errno::ENOENT, Errno::ESRCH
+        # someone else did our job for us
+      end
     end
   end
-end
 
-after_fork do |server, worker|
-  ActiveRecord::Base.establish_connection
+  after_fork do |server, worker|
+    ActiveRecord::Base.establish_connection
+  end
 end
